+++ server.js
@@ -1,12 +1,28 @@
+const mysql = require("mysql2");
+
+require('dotenv').config()
+
 var figlet = require("figlet");
 
-figlet(`Employee    Manager`,
+const db = mysql.createConnection(
+    {
+        host: process.env.DB_HOST,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME
 
-{
-    horizontalLayout: "fitted",
-    verticalLayout: "default",
-    whitespaceBreak: true,
-  },
+    },
+    console.log(`Connected to the hris_db database.`)
+);
+
+figlet(
+    `H R I S`,
+
+    {
+        horizontalLayout: "fitted",
+        verticalLayout: "default",
+        whitespaceBreak: true,
+    },
     function (err, data) {
         if (err) {
             console.log("Unexpected Error");
@@ -14,4 +30,7 @@ figlet(`Employee    Manager`,
             return;
         }
         console.log(data);
-    });
+    }
+);
+
+db.end();