+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 3. Initialised NPM & installed Inquirier and MySql2
 4. Installed figlet & inquirer packages
 5. Created server.js & .env files
+6. Installed dotenv package to move db connection data to .env file
+7. Created db schema for HRIS (emploee) database
+8. Created mysql2 connect/disconnect & tested